+++ src/bl64-tf.bash
@@ -137,9 +137,9 @@ function bl64_tf_run_tofu() {
   bl64_dbg_lib_show_vars 'TF_LOG' 'TF_LOG_PATH'
 
   bl64_dbg_lib_trace_start
-  # shellcheck disable=SC2086
   "$BL64_TF_CMD_TOFU" \
-    "$@"
+    "$@" \
+    -consolidate-errors
   bl64_dbg_lib_trace_stop
 }
 
@@ -157,17 +157,25 @@ function bl64_tf_run_tofu() {
 function _bl64_tf_harden_tofu() {
   bl64_dbg_lib_show_function
 
-  bl64_dbg_lib_show_info 'unset inherited TF_* shell variables'
+  bl64_dbg_lib_show_info 'unset inherited shell variables'
   bl64_dbg_lib_trace_start
+  unset TF_CLI_ARGS
   unset TF_CLI_CONFIG_FILE
   unset TF_DATA_DIR
+  unset TF_ENCRYPTION
   unset TF_IN_AUTOMATION
   unset TF_INPUT
   unset TF_LOG
+  unset TF_LOG_CORE
   unset TF_LOG_PATH
+  unset TF_LOG_PROVIDER
   unset TF_PLUGIN_CACHE_DIR
+  unset TF_PROVIDER_DOWNLOAD_RETRY
   unset TF_REGISTRY_CLIENT_TIMEOUT
   unset TF_REGISTRY_DISCOVERY_RETRY
+  unset TF_STATE_PERSIST_INTERVAL
+  unset TF_WORKSPACE
+  unset TOFU_CPU_PROFILE
   bl64_dbg_lib_trace_stop
 
   return 0